Parameter & channel fetching, separate components
