+++ frontend/app/actions/REST.js
@@ -73,7 +73,21 @@ export default {
         }
     },
 
+    getChannels : function() {
+        return function(dispatch) {
+            makeQuery(dispatch, 'Channels', '/en/api/channels/');
+        }
+    },
+
     /* disabled until proper backend filter */
+    /*
+    getParameters : function() {
+        return function(dispatch) {
+            makeQuery(dispatch, 'Parameters', '/en/api/parameters');
+        }
+    },*/
+
+    /* also disabled until proper backend filter */
     /*
     getMeasurements : function() {
         return function(dispatch) {
@@ -85,6 +99,49 @@ export default {
         }
     }*/
 
+    /* used until backend fix */
+    getParameters : function(channels) {
+        return function(dispatch) {
+            dispatch({
+                type: Enum['Parameters' + RESTState.pending],
+                payload: true
+            });
+
+            let promises = new Array();
+            let parameters = new Array();
+
+            axios.get('/en/api/channels').then(function(response) {
+                if(Array.isArray(response.data.results) && response.data.results.length > 0) {
+                    response.data.results.forEach(function(channel) {
+                        promises.push(axios.get('/en/api/parameters', {
+                            params: {
+                                channel: channel.id
+                            }
+                        }));
+                    });
+
+                    axios.all(promises).then(axios.spread(function(...responses) {
+                        responses.forEach(function(response) {
+                            if(Array.isArray(response.data.results) && response.data.results.length > 0) {
+                                /*dispatch({
+                                    type: Enum.PushMeasurement,
+                                    payload: response.data.results[0].id
+                                })*/
+                                parameters.push(response.data.results[0]);
+                            }
+                        });
+                    })).then(function(){
+                        dispatch({
+                            type: Enum['Parameters' + RESTState.completed],
+                            payload: parameters
+                        });
+                    });
+                }
+            });
+        }
+    },
+
+    /* also used until backend fix */
     getMeasurements : function(sessions, params) {
         return function(dispatch) {
             dispatch({
