+++ frontend/app/components/ChannelParameterPicker.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import { Col, Button, FormGroup, ProgressBar } from 'react-bootstrap';
+
+export default class ChannelParameterPicker extends Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (<div>chan params</div>);
+    }
+}
